backend/api_gateway/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
python-dotenv==1.0.1
drf-yasg==1.21.7
